Time: 16 ms (24.20%) | Memory: 59.4 MB (87.17%) - LeetSync

+++ 2006-find-the-student-that-will-replace-the-chalk/find-the-student-that-will-replace-the-chalk.java
@@ -1,7 +1,7 @@
 class Solution {
     public int chalkReplacer(int[] chalk, int k) {
         Long sum = Arrays.stream(chalk).mapToLong(n -> Long.valueOf(n)).sum();
-        int remainder = (int) (k % sum);
+        Long remainder = k % sum;
         
         for(int i = 0; i < chalk.length; i++) {
             if(chalk[i] > remainder) {
